requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
chromadb
sentence-transformers
pymongo
pandas
requests
python-multipart
langgraph=0.3.12
streamlit